Add loadDeviceInfoIfNeeded with completion
Using the existing available APIs, it is possible to check and fetch
the device info, however each usage requires first checking isInfoLoaded
and then optionally making the request to fetch, resulting in duplicate
logic branches. This change is non-breaking and merely makes the
request in the case that the info has not been loaded yet, invoking the
completion block after either the info is loaded or an error occurs.