Replace YapDatabase archiver implementation
Swap the YapDatabase persistence backend with pure NSKeyedArchiver. This
change should mitigate a number of bugs relating to background access as
well as remove weak frameworks from dependencies, so bitcode can be
enabled.

The new implementation makes no efforts to back up the existing store,
so the cache is dropped immediately.